refactor: extract Haversine formula to utils/geo.js
