+++ src/utils/geo.js
@@ -1,0 +1,24 @@
+/**
+ * Calculates the Haversine distance between two geographic coordinates.
+ * @param {{ lat: number, lng: number }} coordA
+ * @param {{ lat: number, lng: number }} coordB
+ * @returns {{ distanceMeters: number, distanceMiles: number }}
+ */
+export function calculateDistance(coordA, coordB) {
+  const toRad = (deg) => (deg * Math.PI) / 180;
+  const earthRadiusMeters = 6371000;
+  const dLat = toRad(coordB.lat - coordA.lat);
+  const dLng = toRad(coordB.lng - coordA.lng);
+  const a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(toRad(coordA.lat)) *
+      Math.cos(toRad(coordB.lat)) *
+      Math.sin(dLng / 2) *
+      Math.sin(dLng / 2);
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+  const distanceMeters = earthRadiusMeters * c;
+  return {
+    distanceMeters,
+    distanceMiles: distanceMeters / 1609.344,
+  };
+}
